+++ Core/Core/API/APIActivity.swift
@@ -41,3 +41,17 @@ public struct APIActivity: Codable {
     let course_id: ID?
     let group_id: ID?
 }
+
+public struct GetActivitiesRequest: APIRequestable {
+    public typealias Response = APIActivity
+
+    public var path: String {
+        return "/users/self/activity_stream"
+    }
+
+    public var query: [APIQueryItem] {
+        return [
+            .value("per_page", "99"),
+        ]
+    }
+}
